requirements.txt
google-play-scraper
pandas
python-dateutil
google-generativeai
python-dotenv
